refactor: 拆分expand界面到不同模块
